Refactor ReplacementTrap and StormSignal components: Remove useNavigate and replace with useBackToCaseStudies hook for improved navigation handling. Streamline back navigation logic to enhance code maintainability.
